+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#269] Functionality to create service accounts after their dependant is already installed
+  - This is necessary as service accounts will be skipped if an optional dependency is not installed.
+    If the dependency is installed later, this will ensure that the service account will be created.
+  - This works by reconciling the dependent dogu.
 
 ## [v3.14.0] - 2025-10-13
 ### Changed